MAPDN/pyproject.toml
[project]
name = "mapdn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.7, <3.8"
dependencies = [
    "gymnasium>=0.26.3",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "imageio>=2.9.0",
    "joblib==0.17.0",
    "matplotlib==2.2.5",
    # "mpi4py==3.0.3",
    "numpy==1.19.5",
    "numpy-stl==2.16.0",
    "omegaconf>=2.3.0",
    "pandapower==2.7.0",
    "pandas==1.1.3",
    "plotly==5.3.1",
    "pyglet==2.0.dev0",
    "pyyaml==6.0.1",
    "rich>=0.3.3",
    "ruff>=0.7.3",
    "six==1.15.0",
    "tensorboard==2.1.1",
    "tensorboard-data-server==0.6.0",
    "tensorboard-plugin-wit==1.7.0",
    "tensorboardx==2.2",
    "torch==1.10.0+cu113",
    "wandb>=0.13.5",
    "numba>=0.56.4",
    "xvfbwrapper>=0.2.9",
    "tqdm>=4.67.0",
    "python-dotenv>=0.21.1",
    "hydra-joblib-launcher>=1.2.0",
]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "cstcloud"
url = "https://pypi.cstcloud.cn/simple"
default = true

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl"
explicit = true

# [tool.uv.workspace]
# members = [
#   "packages/*"
# ]